Add files via upload
This is a login page is made  using html and css. This login page provides users with secure access to their accounts. The design is clean and intuitive, featuring a simple form that asks for a username or email and a password.
"Forgot Password?" link: Allows users to reset their password if needed.
Login Button: A clearly visible button that submits the form for authentication